Trie snapshots
