+++ packages/client-chains/package.json
@@ -32,15 +32,15 @@
     "@polkadot/client-db": "^0.15.4",
     "@polkadot/client-runtime": "^0.15.4",
     "@polkadot/client-wasm": "^0.15.4",
-    "@polkadot/storage": "^0.28.25",
-    "@polkadot/util": "^0.29.21",
-    "@polkadot/util-crypto": "^0.29.21",
-    "@polkadot/util-keyring": "^0.29.21"
+    "@polkadot/storage": "^0.28.27",
+    "@polkadot/util": "^0.29.22",
+    "@polkadot/util-crypto": "^0.29.22",
+    "@polkadot/util-keyring": "^0.29.22"
   },
   "devDependencies": {
     "@polkadot/client": "^0.15.4",
-    "@polkadot/extrinsics": "^0.28.25",
-    "@polkadot/primitives": "^0.28.25",
-    "@polkadot/storage": "^0.28.25"
+    "@polkadot/extrinsics": "^0.28.27",
+    "@polkadot/primitives": "^0.28.27",
+    "@polkadot/storage": "^0.28.27"
   }
 }
