+++ packages/client-db/package.json
@@ -29,16 +29,16 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-db#readme",
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
-    "@polkadot/db": "^0.29.21",
-    "@polkadot/storage": "^0.28.25",
-    "@polkadot/trie-db": "^0.29.21",
-    "@polkadot/trie-hash": "^0.29.21",
-    "@polkadot/util": "^0.29.21",
-    "@polkadot/util-crypto": "^0.29.21",
+    "@polkadot/db": "^0.29.22",
+    "@polkadot/storage": "^0.28.27",
+    "@polkadot/trie-db": "^0.29.22",
+    "@polkadot/trie-hash": "^0.29.22",
+    "@polkadot/util": "^0.29.22",
+    "@polkadot/util-crypto": "^0.29.22",
     "@types/mkdirp": "^0.5.2",
     "mkdirp": "^0.5.1"
   },
   "devDependencies": {
-    "@polkadot/storage": "^0.28.25"
+    "@polkadot/storage": "^0.28.27"
   }
 }
