+++ packages/client-p2p-messages/package.json
@@ -30,8 +30,8 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-p2p-messages#readme",
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
-    "@polkadot/jsonrpc": "^0.28.25",
-    "@polkadot/primitives": "^0.28.25",
-    "@polkadot/util": "^0.29.21"
+    "@polkadot/jsonrpc": "^0.28.27",
+    "@polkadot/primitives": "^0.28.27",
+    "@polkadot/util": "^0.29.22"
   }
 }
