+++ packages/client-p2p/package.json
@@ -31,8 +31,8 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
     "@polkadot/client-p2p-messages": "^0.15.4",
-    "@polkadot/primitives": "^0.28.25",
-    "@polkadot/util": "^0.29.21",
+    "@polkadot/primitives": "^0.28.27",
+    "@polkadot/util": "^0.29.22",
     "eventemitter3": "^3.1.0",
     "libp2p": "^0.23.1",
     "libp2p-bootstrap": "^0.9.3",
