+++ packages/client-rpc-handlers/package.json
@@ -26,8 +26,8 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-rpc-handlers#readme",
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
-    "@polkadot/jsonrpc": "^0.28.25",
-    "@polkadot/primitives": "^0.28.25",
-    "@polkadot/util": "^0.29.21"
+    "@polkadot/jsonrpc": "^0.28.27",
+    "@polkadot/primitives": "^0.28.27",
+    "@polkadot/util": "^0.29.22"
   }
 }
