+++ packages/client-rpc/package.json
@@ -29,13 +29,13 @@
   },
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-rpc#readme",
   "devDependencies": {
-    "@polkadot/api-provider": "^0.28.25",
+    "@polkadot/api-provider": "^0.28.27",
     "@polkadot/client": "^0.15.4"
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
     "@polkadot/client-rpc-handlers": "^0.15.4",
-    "@polkadot/util": "^0.29.21",
+    "@polkadot/util": "^0.29.22",
     "@types/co-body": "^0.0.3",
     "@types/koa": "^2.0.46",
     "@types/koa-route": "^3.2.4",
