+++ packages/client-runtime/package.json
@@ -31,7 +31,7 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
     "@polkadot/client-db": "^0.15.4",
-    "@polkadot/util": "^0.29.21",
-    "@polkadot/util-crypto": "^0.29.21"
+    "@polkadot/util": "^0.29.22",
+    "@polkadot/util-crypto": "^0.29.22"
   }
 }
