+++ packages/client-telemetry/package.json
@@ -30,7 +30,7 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-telemetry#readme",
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
-    "@polkadot/util": "^0.29.21",
+    "@polkadot/util": "^0.29.22",
     "websocket": "^1.0.26"
   }
 }
