+++ packages/client-wasm/package.json
@@ -31,8 +31,8 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0-rc.1",
     "@polkadot/client-runtime": "^0.15.4",
-    "@polkadot/primitives": "^0.28.25",
-    "@polkadot/util": "^0.29.21",
+    "@polkadot/primitives": "^0.28.27",
+    "@polkadot/util": "^0.29.22",
     "@types/webassembly-js-api": "^0.0.1"
   }
 }
