+++ packages/client/package.json
@@ -39,9 +39,9 @@
     "@polkadot/client-rpc": "^0.15.4",
     "@polkadot/client-telemetry": "^0.15.4",
     "@polkadot/client-wasm": "^0.15.4",
-    "@polkadot/jsonrpc": "^0.28.25",
-    "@polkadot/primitives": "^0.28.25",
-    "@polkadot/util": "^0.29.21",
+    "@polkadot/jsonrpc": "^0.28.27",
+    "@polkadot/primitives": "^0.28.27",
+    "@polkadot/util": "^0.29.22",
     "@types/package-json": "^4.0.1",
     "@types/semver-compare": "^1.0.0",
     "@types/yargs": "^11.1.1",
